Improve rsync efficiency

+++ scripts/update_sifts_via_rsync_worker.sh
@@ -55,10 +55,23 @@ do
 
  if [ $proc_id_mod = $PROC_ID ] ; then
 
-  rsync -a --delete $SRC_DIR/$div_dir/ $DST_DIR/$div_dir
+  lock_file=$DST_DIR/$div_dir/.lock
+  done_file=$DST_DIR/$div_dir/.done
 
-  if [ $proc_id_mod -eq 0 ] ; then
-   echo -e -n "\rDone "$((proc_id + 1)) of $TOTAL ...
+  if [ ! -e $lock_file ] && [ ! -e $done_file ] ; then
+
+   touch $lock_file
+
+   rsync -a --delete $SRC_DIR/$div_dir/ $DST_DIR/$div_dir
+
+   touch $done_file
+
+   if [ $proc_id_mod -eq 0 ] ; then
+    echo -e -n "\rDone "$((proc_id + 1)) of $TOTAL ...
+   fi
+
+   rm -f $lock_file
+
   fi
 
  fi
@@ -66,4 +79,34 @@ do
  let proc_id++
 
 done < $DIR_LIST
+
+proc_id=0
+
+while read div_dir
+do
+
+ proc_id_mod=$(($proc_id % $MAXPROCS))
+
+ if [ $proc_id_mod = $PROC_ID ] ; then
+
+  lock_file=$DST_DIR/$div_dir/.lock
+  done_file=$DST_DIR/$div_dir/.done
+
+  if [ ! -e $lock_file ] && [ ! -e $done_file ] ; then
+
+   touch $lock_file
+
+   rsync -a --delete $SRC_DIR/$div_dir/ $DST_DIR/$div_dir
+
+   touch $done_file
+
+   rm -f $lock_file
+
+  fi
+
+ fi
+
+ let proc_id++
+
+done < $DIR_LIST~
